services/user-management-service/requirements.txt
fastapi
fastapi-cors
uvicorn
sqlalchemy
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic[email]
passlib[bcrypt]
pyjwt
